add custom icon buttons
